requirements.txt
azure-search-documents
azure-core
openai>=1.0.0
python-dotenv